+++ CC98.LogOn/CC98.LogOn/wwwroot/lib/i18next-browser-languagedetector/i18nextBrowserLanguageDetector.js
@@ -162,6 +162,44 @@
     }
   };
 
+  var path = {
+    name: 'path',
+
+    lookup: function lookup(options) {
+      var found = void 0;
+      if (typeof window !== 'undefined') {
+        var language = window.location.pathname.match(/\/([a-zA-Z-]*)/g);
+        if (language instanceof Array) {
+          if (typeof options.lookupFromUrlIndex === 'number') {
+            found = language[options.lookupFromPathIndex].replace('/', '');
+          } else {
+            found = language[0].replace('/', '');
+          }
+        }
+      }
+      return found;
+    }
+  };
+
+  var subdomain = {
+    name: 'subdomain',
+
+    lookup: function lookup(options) {
+      var found = void 0;
+      if (typeof window !== 'undefined') {
+        var language = window.location.pathname.match(/(?:http[s]*\:\/\/)*(.*?)\.(?=[^\/]*\..{2,5})/gi);
+        if (language instanceof Array) {
+          if (typeof options.lookupFromSubdomainIndex === 'number') {
+            found = language[options.lookupFromSubdomainIndex].replace('http://', '').replace('https://', '').replace('.', '');
+          } else {
+            found = language[0].replace('http://', '').replace('https://', '').replace('.', '');
+          }
+        }
+      }
+      return found;
+    }
+  };
+
   var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
 
   function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
@@ -208,6 +246,8 @@
         this.addDetector(localStorage);
         this.addDetector(navigator$1);
         this.addDetector(htmlTag);
+        this.addDetector(path);
+        this.addDetector(subdomain);
       }
     }, {
       key: 'addDetector',
@@ -237,7 +277,19 @@
           if (_this.services.languageUtils.isWhitelisted(cleanedLng)) found = cleanedLng;
         });
 
-        return found || this.i18nOptions.fallbackLng[0];
+        if (!found) {
+          var fallbacks = this.i18nOptions.fallbackLng;
+          if (typeof fallbacks === 'string') fallbacks = [fallbacks];
+          if (!fallbacks) fallbacks = [];
+
+          if (Object.prototype.toString.apply(fallbacks) === '[object Array]') {
+            found = fallbacks[0];
+          } else {
+            found = fallbacks[0] || fallbacks.default && fallbacks.default[0];
+          }
+        };
+
+        return found;
       }
     }, {
       key: 'cacheUserLanguage',