+++ CC98.LogOn/CC98.LogOn/wwwroot/lib/i18next-browser-languagedetector/CHANGELOG.md
@@ -1,3 +1,9 @@
+### v2.2.0
+- add detector for path and subdomain [PR153](https://github.com/i18next/i18next-browser-languageDetector/pull/153) and [PR152](https://github.com/i18next/i18next-browser-languageDetector/pull/152)
+
+### v2.1.1
+- support for fallback language in form of object [151](https://github.com/i18next/i18next-browser-languageDetector/issues/151)
+
 ### v2.1.0
 - add .js for browser import implementation [PR147](https://github.com/i18next/i18next-browser-languageDetector/pull/147)